modules created

+++ psmnet/models/stackhourglass.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
 
-from submodule import *
+from .submodule import *
 import torch
 import torch.nn as nn
 import torch.utils.data
@@ -113,10 +113,10 @@ class PSMNet(nn.Module):
 
         # matching
         cost = Variable(
-            torch.FloatTensor(refimg_fea.size()[0], refimg_fea.size()[1] * 2, self.maxdisp / 4, refimg_fea.size()[2],
+            torch.FloatTensor(refimg_fea.size()[0], refimg_fea.size()[1] * 2, int(self.maxdisp / 4), refimg_fea.size()[2],
                               refimg_fea.size()[3]).zero_()).cuda()
 
-        for i in range(self.maxdisp / 4):
+        for i in range(int(self.maxdisp / 4) ):
             if i > 0:
                 cost[:, :refimg_fea.size()[1], i, :, i:] = refimg_fea[:, :, :, i:]
                 cost[:, refimg_fea.size()[1]:, i, :, i:] = targetimg_fea[:, :, :, :-i]
